discord-bot/requirements.txt
discord.py
psycopg2-binary   # For PostgreSQL later
